+++ src/pages/opplysninger/[side].tsx
@@ -10,7 +10,7 @@ import BestattUtdanning from '../../components/skjema/utdanning-bestatt';
 import Helseproblemer from '../../components/skjema/helseproblemer';
 import AndreProblemer from '../../components/skjema/andre-problemer';
 import OppsummeringUtenPlikter from '../../components/skjema/oppsummering/oppsummering-uten-plikter';
-import { beregnNavigering } from '../../lib/standard-registrering-uten-plikter-tilstandsmaskin';
+import { beregnNavigering } from '../../lib/standard-registrering-tilstandsmaskin';
 import { SkjemaSide, SkjemaState, visSisteStilling } from '../../model/skjema';
 import { SkjemaAction } from '../../lib/skjema-state';
 import SisteStilling from '../../components/skjema/siste-jobb/siste-stilling';
